// types/contenido.ts
export type NosotrosContenido = {
  heroLabel: string;
  heroTitle: string;
  heroDescription: string;

  focusLabel: string;
  focusTitle: string;
  focusParagraphOne: string;
  focusParagraphTwo: string;

  pillarOneTitle: string;
  pillarOneDescription: string;

  pillarTwoTitle: string;
  pillarTwoDescription: string;

  pillarThreeTitle: string;
  pillarThreeDescription: string;

  pillarFourTitle: string;
  pillarFourDescription: string;

  teamLabel: string;
  teamTitle: string;
  teamDescription: string;

  updatedAt?: unknown;
};

export type InicioContenido = {
  heroLabel: string;
  heroTitle: string;
  heroDescription: string;
  primaryButtonLabel: string;
  primaryButtonHref: string;
  secondaryButtonLabel: string;
  secondaryButtonHref: string;

  servicesLabel: string;
  servicesTitle: string;
  servicesDescription: string;

  studiesLabel: string;
  studiesTitle: string;
  studiesDescription: string;

  sectorsLabel: string;
  sectorsTitle: string;
  sectorsDescription: string;

  contactLabel: string;
  contactTitle: string;
  contactDescription: string;

  updatedAt?: unknown;
};